3-serverless-microservice-stack/requirements.txt
aws_cdk.core==1.43.0
aws_cdk.aws_apigateway==1.43.0
aws_cdk.aws_lambda==1.43.0
aws_cdk.aws_s3==1.43.0
aws_cdk.aws_s3_deployment==1.43.0
aws_cdk.aws_dynamodb==1.43.0
aws-cdk.aws_cloudformation==1.43.0
aws-cdk.aws_events==1.43.0
aws-cdk.aws_events_targets==1.43.0
